requirements.txt
aiocsv==1.2.3
aiofiles~=22.1.0
asyncio==3.4.3
asyncpg==0.27.0
blinker~=1.4
certifi~=2022.9.24
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.6.2
h11~=0.12.0
h2~=3.2.0
hpack~=3.0.0
hypercorn==0.14.3
hyperframe~=5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe~=2.1.1
priority==2.0.0
quart==0.18.3
toml==0.10.2
urllib3~=1.26.12
Werkzeug~=2.2.2
wsproto==1.2.0

typing~=3.7.4.3
pytest~=7.1.3
pip~=22.0.2
emoji~=2.1.0
attrs~=22.1.0
wheel~=0.37.1
Brotli~=1.0.9
ipaddress~=1.0.23
cryptography~=3.4.8
py~=1.11.0
lxml~=4.8.0
Pillow~=9.0.1
keyring~=23.5.0
decorators~=2.0.7
websocket-client~=1.2.3
Flask~=2.1.3
six~=1.16.0
setuptools~=65.5.0
future~=0.18.2
aiohttp~=3.8.3
requests~=2.28.1
python-dotenv~=0.21.0
httpx~=0.23.0